Promoted ring envelope parsing code to its own package moving it out of experiments.
