+++ GUI/reportgeneration/report_generation_controller.py
@@ -4,8 +4,6 @@ import time
 import traceback
 import tkinter as tk
 from tkinter import ttk
-from tkinter import filedialog
-import pandas as pd
 from tkinter.messagebox import showinfo
 
 from GUI.listbox_log_handler import ListboxHandler
@@ -47,11 +45,11 @@ class ReportGenerationController():
 
         import threading
         import LoadTournamentTable
-        import experiments.ringevelopeparser.ProcessRegistrationsWithRingEnvelopeData as ProcessRegistrationsWithRingEnvelopeData
+        import ringevelopeparser.ProcessRegistrationsWithRingEnvelopeData as ProcessRegistrationsWithRingEnvelopeData
         try:
             ltt = LoadTournamentTable.LoadTournamentTable()
             # worker_thread=threading.Thread(target=ltt.process_tournament_table, args=[self.app_container.input_data_filename,self.app_container.database,self.app_container.tournament_output_folder_path])
-            worker_thread = threading.Thread(target=ProcessRegistrationsWithRingEnvelopeData.process_registrations_with_ring_envelope_data, args=[self.app_container.ring_envelope_database_filename,self.app_container.input_data_filename, self.app_container.database,self.app_container.tournament_output_folder_path])
+            worker_thread = threading.Thread(target=ProcessRegistrationsWithRingEnvelopeData.process_registrations_with_ring_envelope_data, args=[self.app_container.ring_envelope_database_filename, self.app_container.input_data_filename, self.app_container.database, self.app_container.tournament_output_folder_path])
             logging.info("go")
             worker_thread.start()
             # self.app_container.after(500,lambda:x.start())
